test: name the attribution tests after the modules they test, and refresh stats
A _test.go exists only under the name of a module it tests, which
make check-test-file-names gates: the three attribution_test.go files named no
module. Each moves into the test file of the module whose guard it drives, which
is where a reader would look for it anyway: the resource registrar, the prompt
registrar, and the completion handler.

The generated README stats and testing reference follow.
